Integrate TextSource with preprocessor functions
The previous preprocess methods in TextSource have been moved to preprocess. The naming convention for the different data files has set fixed (i.e. we have three files for the data, the raw (.csv), the parsed stories in a list (.clean) and the preprocessed file (.processed)). Some other minor fixes the integration.